Refactor font components and improve code organization
- Removed the `fitText` function and related types from `utils.ts`.
- Consolidated font style logic into `getFontStyles` in `use-local-font-query.ts`.
- Updated `FontGlyphPanel`, `FontMetaCard`, and `FontPreviewPanel` to utilize the new `getFontStyles` function.
- Cleaned up unused imports and components, including the removal of the `rive.tsx` file.
- Enhanced the `ShortcutsDialog` component to export as a named function.
- Made minor adjustments to CSS styles for consistency.
