+++ src/app/globals.css
@@ -4,8 +4,7 @@
   --background: oklch(0.9252 0.0094 62.58);
   --foreground:oklch(0.1559 0.0074 102.71);
   --accent: #FFA112;
-  --border-dashed: oklch(0.25 0.0074 102.71 / 1); /* darker than foreground/10 */
-
+  --border-dashed: oklch(0.25 0.0074 102.71 / 1); 
 }
 
 @theme inline {
@@ -17,6 +16,7 @@
   --color-dashed: var(--border-dashed);
   
 }
+
 *::selection {
   background-color: color-mix(in oklch, var(--accent), white 10%);
   color: var(--foreground);
@@ -34,7 +34,7 @@
     --background: oklch(0.1559 0.0074 102.71);
     --foreground: oklch(0.9252 0.0094 62.58);
     --accent: #FFA112;
-    --border-dashed: oklch(0.7 0.0094 62.58 / 1); /* darker for dark mode */
+    --border-dashed: oklch(0.7 0.0094 62.58 / 1); 
 
   }
 }
@@ -43,8 +43,6 @@ body {
   background: var(--background);
   color: var(--foreground);
 }
-
-
 
 .root {
   isolation: isolate;
@@ -55,8 +53,4 @@ body {
   overscroll-behavior: none;
 }
 
-.font-card:focus-visible {
-  outline: none;
-  z-index: 1;
-}
