+++ src/components/animated-rive-logo.tsx
@@ -3,7 +3,8 @@
 import { useRive, Layout, Fit, Alignment } from "@rive-app/react-webgl2";
 import { useRef } from "react";
 
-const RiveImpl = ({ src }: { src: string }) => {
+export const AnimatedRiveLogo = ({ src }: { src: string }) => {
+  const ref = useRef<HTMLDivElement>(null);
   const { RiveComponent } = useRive({
     src: src,
     stateMachines: "State Machine 1",
@@ -14,17 +15,10 @@ const RiveImpl = ({ src }: { src: string }) => {
     }),
     autoplay: true,
   });
-  
-
-  return <RiveComponent />;
-};
-
-export const Rive = ({ src }: { src: string }) => {
-  const ref = useRef<HTMLDivElement>(null);
 
   return (
     <div className="w-full h-full" ref={ref}>
-      <RiveImpl src={src} />
+      <RiveComponent />
     </div>
   );
 };
